Phase 5: type annotation coverage improvement (61.9% -> 85%+)
Add explicit type annotations to reduce mypy strict errors from 825 to
110 across the emulator subpackage:

- types.py: declare BabbageNumber.value and _overflow_flag as typed
  instance attributes; fix to_decimal() return type
- structural.py: float() cast on power expressions to suppress no-any-return
- tribology.py: float() cast on EHL film thickness return
- carry.py, barrels.py, machine.py, card_reader.py, timing.py: add
  missing __init__ return annotations and fix Returning Any errors
- timing.py: type BARREL_TIMING_MAP as dict[str, dict[str, int|list[str]]]
  to enable correct type narrowing in BarrelTimingBridge methods
- pascaline.py, leibniz_reckoner.py, curta.py, napiers_bones.py,
  enigma.py, antikythera.py: add missing -> None / -> int / -> str
  return annotations throughout
- quipu.py, clay_tokens.py: add explicit var annotations for _records,
  _tokens, _impression to satisfy var-annotated checker

All 1135 unit tests pass after these changes. No behavior changes.

+++ backend/src/emulator/structural.py
@@ -17,9 +17,6 @@ Acceptance criteria:
 from __future__ import annotations
 
 import math
-from dataclasses import dataclass
-from typing import Optional, Tuple
-
 
 # ---------------------------------------------------------------------------
 # Shaft Deflection Analysis (Euler-Bernoulli, Shigley Ch.4)
@@ -151,7 +148,7 @@ class FatigueAnalysis:
             "forged": (272.0, -0.995),
         }
         a, b = factors.get(finish, factors["machined"])
-        return a * Su_MPa**b
+        return float(a * Su_MPa**b)
 
     @staticmethod
     def size_factor_kb(diameter_mm: float) -> float:
@@ -163,9 +160,9 @@ class FatigueAnalysis:
         if diameter_mm <= 2.79:
             return 1.0
         elif diameter_mm <= 51.0:
-            return 1.24 * diameter_mm**(-0.107)
+            return float(1.24 * diameter_mm**(-0.107))
         elif diameter_mm <= 254.0:
-            return 1.51 * diameter_mm**(-0.157)
+            return float(1.51 * diameter_mm**(-0.157))
         else:
             return 0.6
 
@@ -229,7 +226,7 @@ class FatigueAnalysis:
         a = s_low / (n_low ** (-b))
         if b == 0:
             return 1e6
-        return (stress_amplitude_MPa / a) ** (1.0 / (-b))
+        return float((stress_amplitude_MPa / a) ** (1.0 / (-b)))
 
 
 # ---------------------------------------------------------------------------
@@ -399,7 +396,7 @@ class StressConcentration:
         else:
             A_coeff, b_exp = 1.04, -0.26
 
-        K_t = A_coeff * ratio_rd ** b_exp
+        K_t = float(A_coeff * ratio_rd ** b_exp)
         return max(1.0, K_t)
 
     @staticmethod
@@ -439,7 +436,7 @@ class DynamicLoadFactor:
         if V == 0.0:
             return 1.0
 
-        K_v = ((A + math.sqrt(V)) / A) ** B
+        K_v = float(((A + math.sqrt(V)) / A) ** B)
         return max(1.0, K_v)
 
     @staticmethod
